backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
boto3
ultralytics
Pillow
numpy
python-dotenv
mangum